Move canUser to the appropriate location
